# swagger/Trainingsplanner-postgres.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "laplanner",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/appointments": {
      "get": {
        "tags": [
          "TrainingsAppointment"
        ],
        "operationId": "TrainingsAppointment_GetAllAppointments",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingsAppointmentDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "TrainingsAppointment"
        ],
        "operationId": "TrainingsAppointment_CreateAppointment",
        "requestBody": {
          "x-name": "trainingsAppointment",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingsAppointmentDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsAppointmentDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "TrainingsAppointment"
        ],
        "operationId": "TrainingsAppointment_UpdateAppointment",
        "requestBody": {
          "x-name": "trainingsAppointment",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingsAppointmentDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsAppointmentDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "TrainingsAppointment"
        ],
        "operationId": "TrainingsAppointment_DeleteAppointment",
        "requestBody": {
          "x-name": "trainingsAppointment",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingsAppointmentDto"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/{id}": {
      "get": {
        "tags": [
          "TrainingsAppointment"
        ],
        "operationId": "TrainingsAppointment_GetAppointmentById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsAppointmentDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/{id}/full": {
      "get": {
        "tags": [
          "TrainingsAppointment"
        ],
        "operationId": "TrainingsAppointment_GetFullAppointmentById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingsAppointmentDto"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/byUser/{userId}": {
      "get": {
        "tags": [
          "TrainingsAppointment"
        ],
        "summary": "Returns the TrainingsAppointments of spcified User",
        "operationId": "TrainingsAppointment_GetTrainingsAppointmentsByUserId",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "The TrainingsModule",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingsAppointmentDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/{id}/modules": {
      "get": {
        "tags": [
          "TrainingsAppointment"
        ],
        "operationId": "TrainingsAppointment_GetModulesByAppointmentId",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingsModuleDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/{trainingsAppointmentId}": {
      "post": {
        "tags": [
          "TrainingsAppointment"
        ],
        "operationId": "TrainingsAppointment_AddModuleToAppointment",
        "parameters": [
          {
            "name": "trainingsAppointmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "requestBody": {
          "x-name": "moduleDtos",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/TrainingsModuleDto"
                }
              }
            }
          },
          "required": true,
          "x-position": 2
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingsAppointmentTrainingsModuleDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/appointments/{appointmentId}/modules/{moduleId}": {
      "delete": {
        "tags": [
          "TrainingsAppointment"
        ],
        "operationId": "TrainingsAppointment_DeleteAppointmentWithModule",
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "moduleId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "202": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/modules/appointments/{id}": {
      "get": {
        "tags": [
          "TrainingsModule"
        ],
        "summary": "Returns all TrainingsModules that are part of the given appointment id",
        "operationId": "TrainingsModule_GetTrainingsModuleByAppointmentId",
        "parameters": [
          {
            "name": "trainingsAppointmentId",
            "in": "query",
            "description": "Id of the needed TrainingsAppointment",
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          },
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "200": {
            "description": "The TrainingsModule",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TrainingsModuleDto"
                  }
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "TrainingsAppointmentDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "date-time"
          },
          "startTime": {
            "type": "string",
            "format": "date-time"
          },
          "trainingsGroupId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsGroup": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsGroup"
              }
            ]
          },
          "trainingsAppointmentsTrainingsModules": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsAppointmentTrainingsModule"
            }
          }
        }
      },
      "TrainingsGroup": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "trainingsGroupsApplicationUsers": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsGroupApplicationUser"
            }
          },
          "trainingsAppointments": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsAppointment"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "TrainingsAppointmentTrainingsModule": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "trainingsModuleId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModule": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsModule"
              }
            ]
          },
          "trainingsAppointmentId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsAppointment": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsAppointment"
              }
            ]
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "TrainingsDifficulty": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "UNKNOWN",
          "EASY",
          "MEDIUM",
          "HARD",
          "IMPOSSIBLE"
        ],
        "enum": [
          0,
          10,
          20,
          30,
          40
        ]
      },
      "TrainingsModuleTrainingsExercise": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "trainingsModuleId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModule": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsModule"
              }
            ]
          },
          "trainingsExerciesId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsExercise": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsExercise"
              }
            ]
          },
          "position": {
            "type": "integer",
            "format": "int32"
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "TrainingsModuleTrainingsModuleTag": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "trainingsModuleId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModule": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsModule"
              }
            ]
          },
          "trainingsModuleTagId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModuleTag": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsModuleTag"
              }
            ]
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": {
          "nullable": true
        },
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "TrainingsModuleDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "difficulty": {
            "$ref": "#/components/schemas/TrainingsDifficulty"
          },
          "userId": {
            "type": "string",
            "nullable": true
          },
          "trainingsModulesTrainingsExercises": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsModuleTrainingsExercise"
            }
          },
          "trainingsModulesTrainingsModuleTags": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsModuleTrainingsModuleTag"
            }
          },
          "trainingsAppointmentsTrainingsModules": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/TrainingsAppointmentTrainingsModule"
            }
          },
          "created": {
            "type": "string",
            "format": "date-time"
          },
          "updated": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "TrainingsAppointmentTrainingsModuleDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "trainingsModuleId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsModule": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsModule"
              }
            ]
          },
          "trainingsAppointmentId": {
            "type": "integer",
            "format": "int32"
          },
          "trainingsAppointment": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/TrainingsAppointment"
              }
            ]
          }
        }
      },
...
    }
  }
}